pyproject.toml
[project]
name = "darkbotx"
version = "0.1.0"
description = "A visionless grasping robot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "genesis-world>=0.4.0",
    "mapbox-earcut>=2.0.0",
    "numpy-stl>=3.2.0",
    "urdf-parser-py>=0.0.4",
    "pyqt6>=6.10.2",
    "rsl-rl-lib>=4.0.1",
    "torch>=2.10.0",
    "tensorboard>=2.20.0",
    "wandb>=0.25.0",
    "adafruit-circuitpython-servokit>=1.3.22",
    "adafruit-circuitpython-ads1x15>=3.0.2",
    "tabulate>=0.10.0",
]

# "rpi-lgpio>=0.6" is required to be added but version does not match